+++ BrickController2/BrickController2/PlatformServices/BluetoothLE/IBluetoothLEAdvertiserDevice .cs
@@ -53,10 +53,10 @@ namespace BrickController2.PlatformServices.BluetoothLE
     public interface IBluetoothLEAdvertiserDevice :
             IDisposable
     {
-        Task<bool> StartAdvertiseAsync(AdvertisingInterval advertisingIterval, TxPowerLevel txPowerLevel, int manufacturerId, byte[] rawData);
+        Task<bool> StartAdvertiseAsync(AdvertisingInterval advertisingIterval, TxPowerLevel txPowerLevel, ushort manufacturerId, byte[] rawData);
 
         Task<bool> StopAdvertiseAsync();
 
-        bool ChangeAdvertiseAsync(int manufacturerId, byte[] rawData);
+        bool ChangeAdvertiseAsync(ushort manufacturerId, byte[] rawData);
     }
 }
